Separate event to a new file
